+++ library/gcc/include/sys/timer.h
@@ -66,7 +66,7 @@ static inline uint16_t rtc_get_datetime(uint16_t field) {
 		: "b" (field), "Rah" ((uint8_t) 0x02)
 		: "cc", "memory"
 	);
-	return result;
+	return (uint16_t) result;
 }
 
 void rtc_set_datetime_struct(const datetime_t __far* value);
