+++ library/gcc/src/auto/bank_read_word.c
@@ -8,5 +8,5 @@ uint8_t bank_read_word(uint16_t bank_id, uint16_t offset) {
 		: "b" (bank_id), "d" (offset), "Rah" ((uint8_t) 0x04)
 		: "cc", "memory"
 	);
-	return result;
+	return (uint8_t) result;
 }
