+++ library/gcc/src/auto/key_press_check.c
@@ -8,5 +8,5 @@ uint16_t key_press_check(void) {
 		: "Rah" ((uint8_t) 0x00)
 		: "cc", "memory"
 	);
-	return result;
+	return (uint16_t) result;
 }
